+++ security/provenance/type.c
@@ -39,7 +39,6 @@ static const char RL_STR_MMAP[]                  = "mmap";
 static const char RL_STR_MUNMAP[]                = "munmap";
 static const char RL_STR_SHMDT[]                 = "shmdt";
 static const char RL_STR_LINK[]                  = "link";
-static const char RL_STR_LINK_INODE[]            = "link_inode";
 static const char RL_STR_UNLINK[]                = "unlink";
 static const char RL_STR_SPLICE[]                = "splice";
 static const char RL_STR_SETATTR[]               = "setattr";
@@ -169,8 +168,6 @@ const char* relation_str(uint64_t type)
 		return RL_STR_SHMDT;
 	case RL_LINK:
 		return RL_STR_LINK;
-	case RL_LINK_INODE:
-		return RL_STR_LINK_INODE;
 	case RL_UNLINK:
 		return RL_STR_UNLINK;
 	case RL_SPLICE:
@@ -306,7 +303,6 @@ uint64_t relation_id(const char* str)
 	MATCH_AND_RETURN(str, RL_STR_MUNMAP, RL_MUNMAP);
 	MATCH_AND_RETURN(str, RL_STR_SHMDT, RL_SHMDT);
 	MATCH_AND_RETURN(str, RL_STR_LINK, RL_LINK);
-	MATCH_AND_RETURN(str, RL_STR_LINK_INODE, RL_LINK_INODE);
 	MATCH_AND_RETURN(str, RL_STR_UNLINK, RL_UNLINK);
 	MATCH_AND_RETURN(str, RL_STR_SPLICE, RL_SPLICE);
 	MATCH_AND_RETURN(str, RL_STR_SETATTR, RL_SETATTR);