+++ CHANGES.md
@@ -36,6 +36,7 @@
 
 ### v0.4.2
 ```
+- Delete relation RL_LINK_INODE.
 - Added support for unlink operation (RL_UNLINK)
 - Update to kernel version 4.16.13.
 ```